AI_Chinese_Class/requirements.txt
streamlit
azure-cognitiveservices-speech
firebase-admin
google-generativeai
audio-recorder-streamlit